pyproject.toml
[project]
name = "ai-incubator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.27",
    "langchain-tavily>=0.2.7",
    "langgraph>=0.5.1",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.9",
]
